Add new leadership API

+++ apps/icona/src/main/java/org/onosproject/icona/utils/BitSetIndex.java
@@ -3,8 +3,6 @@ package org.onosproject.icona.utils;
 import static org.slf4j.LoggerFactory.getLogger;
 
 import java.util.BitSet;
-import java.lang.Math;
-
 import org.slf4j.Logger;
 
 public class BitSetIndex {
@@ -17,7 +15,7 @@ public class BitSetIndex {
     public enum IndexType {
 
         MGMT_CHANNEL(32),
-        MPLS_LABEL((int) 0xFFFFF);
+        MPLS_LABEL(0xFFFFF);
         // PSEUDOWIRE_CHANNEL((int) Math.pow(2, 64));
 
         protected Integer value;
